cleaned code

+++ engine/renderer/RaylibRenderer.hpp
@@ -27,7 +27,7 @@ class RaylibRenderer : public Renderer
 
         float get_model_height(int model_id) override;
 
-        void play_selected_animation(int model_id, int anim_num, float frame) override;
+        void play_selected_animation(int model_id, int anim_num, int frame) override;
 
     private:
         Camera3D camera;
