+++ engine/renderer/Renderer.hpp
@@ -28,7 +28,7 @@ class Renderer
 
         virtual float get_model_height(int model_id) = 0;
 
-        virtual void play_selected_animation(int model_id, int anim_num, float frame) = 0;
+        virtual void play_selected_animation(int model_id, int anim_num, int frame) = 0;
 };
 
 std::unique_ptr<Renderer> CreateRenderer();
